change double rating > byte rating

+++ src/main/java/com/fdmgroup/backend_streamhub/videostream/model/Series.java
@@ -25,7 +25,7 @@ public class Series {
     private String description;
     private LocalDate releaseDate;
     private String thumbnailURL;
-    private double rating;
+    private byte rating;
     private String cast;
 
     @OneToMany(mappedBy = "series")
